+++ src/index.css
@@ -1,12 +1,17 @@
 html, body {
-    background-color: #21232A;
-    color: white;
-    font-family: Helvetica, Arial , sans-serif;
+    background-color: #6e2670;
+    color: rgb(20, 117, 173);
+    font-family: oblique , serif , sans-serif;
     font-size: 1.3rem;
     padding: 70px;
+    border: 1px solid rgb(255, 255, 255);
+border-radius: 5px; 
 }
 
 button {
+    font-family: oblique , serif , sans-serif;
+    background-color: rgb(20, 117, 173);
+    color: rgb(3, 3, 0);
     font-size: 1.3rem;
     margin-right: 10px;
     padding: 10px;
